requirements.txt
google-generativeai>=0.8.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
markdown2>=2.4.0
flask>=3.0.0
